move methods pertaining to User over to UserRepository
